feat: create an expo app to demonstrate the lib
